+++ Gemfile
@@ -1,0 +1,11 @@
+source :rubygems
+
+gemspec
+
+group :test do
+  gem 'minitest'
+  gem 'minitest-reporters', '>= 0.5.0'
+  gem 'mocha'
+end
+
+gem 'rake'
