+++ lib/version.rb
@@ -1,0 +1,3 @@
+module Rnvivo
+  VERSION = '0.0.1'
+end
